+++ src/utils/cert_manager/challenge/dns01/route53.rs
@@ -10,68 +10,11 @@ use aws_sdk_route53::{
     },
 };
 use color_eyre::eyre::eyre;
-use instant_acme::{AuthorizationHandle, ChallengeType};
-use reqwest::Client;
-use serde_json::json;
 use tokio::sync::RwLock;
 use tracing::info;
 
-use crate::cert_manager::challenge::{ChallengeError, ChallengeHandler, CleanupFuture};
-
-/// Interface for updating DNS records
-#[async_trait]
-pub trait DnsUpdater: Send + Sync {
-    /// Upsert a DNS record for the given domain
-    async fn upsert_record(&self, domain: &str, value: &str) -> Result<(), ChallengeError>;
-    /// Remove a DNS record for the given domain
-    async fn remove_record(&self, domain: &str, value: &str) -> Result<(), ChallengeError>;
-}
-
-/// Handler for DNS-01 challenges
-pub struct Dns01Handler {
-    dns_updater: Arc<dyn DnsUpdater>,
-}
-
-impl Dns01Handler {
-    pub fn new(dns_updater: impl DnsUpdater + 'static) -> Self {
-        Self {
-            dns_updater: Arc::new(dns_updater),
-        }
-    }
-}
-
-#[async_trait]
-impl ChallengeHandler for Dns01Handler {
-    async fn handle_authorization<'a>(
-        &'a self,
-        authz: &'a mut AuthorizationHandle<'a>,
-    ) -> Result<CleanupFuture, ChallengeError> {
-        let mut challenge = authz
-            .challenge(ChallengeType::Dns01)
-            .ok_or_else(|| ChallengeError::Other(eyre!("No DNS-01 challenge found")))?;
-
-        let digest = challenge.key_authorization().dns_value();
-        let domain = challenge.identifier().to_string();
-
-        // Upsert the DNS record
-        self.dns_updater.upsert_record(&domain, &digest).await?;
-
-        // Signal the server we are ready to respond to the challenge
-        challenge.set_ready().await?;
-
-        let cleanup = {
-            let dns_updater = self.dns_updater.clone();
-            async move { dns_updater.remove_record(&domain, &digest).await }
-        };
-        Ok(CleanupFuture::new(cleanup))
-    }
-}
-
-impl From<instant_acme::Error> for ChallengeError {
-    fn from(err: instant_acme::Error) -> Self {
-        ChallengeError::Other(err.into())
-    }
-}
+use super::DnsUpdater;
+use crate::cert_manager::challenge::ChallengeError;
 
 /// A DNS updater for AWS Route 53
 pub struct AwsRoute53DnsUpdater {
@@ -329,58 +272,6 @@ impl DnsUpdater for AwsRoute53DnsUpdater {
             .await?;
 
         info!("DNS record {record_name} deleted for {domain}");
-        Ok(())
-    }
-}
-
-/// Handler for Pebble DNS (mainly for generating test certificates)
-pub struct PebbleDnsUpdater {
-    client: Client,
-    addr: String,
-}
-
-impl PebbleDnsUpdater {
-    pub fn new(addr: impl Into<String>) -> Self {
-        Self {
-            client: Client::new(),
-            addr: addr.into(),
-        }
-    }
-}
-
-#[async_trait]
-impl DnsUpdater for PebbleDnsUpdater {
-    async fn upsert_record(&self, domain: &str, value: &str) -> Result<(), ChallengeError> {
-        let record_name = format!("_acme-challenge.{domain}.");
-        let url = format!("{}/set-txt", self.addr);
-        let body = json!({"host": record_name, "value": value});
-
-        self.client
-            .post(&url)
-            .body(body.to_string())
-            .send()
-            .await
-            .map_err(|e| ChallengeError::Other(eyre!("Failed to send request: {e}")))?
-            .error_for_status()
-            .map_err(|e| ChallengeError::Other(eyre!("Failed to set TXT record: {e}")))?;
-
-        Ok(())
-    }
-
-    async fn remove_record(&self, domain: &str, _value: &str) -> Result<(), ChallengeError> {
-        let record_name = format!("_acme-challenge.{domain}.");
-        let url = format!("{}/clear-txt", self.addr);
-        let body = json!({"host": record_name});
-
-        self.client
-            .post(&url)
-            .body(body.to_string())
-            .send()
-            .await
-            .map_err(|e| ChallengeError::Other(eyre!("Failed to send request: {e}")))?
-            .error_for_status()
-            .map_err(|e| ChallengeError::Other(eyre!("Failed to clear TXT record: {e}")))?;
-
         Ok(())
     }
 }
